requirements.txt
pandas
numpy
scikit-learn
lightgbm
xgboost
catboost
joblib
pyyaml
pyarrow
fastparquet